camera: fix c++11 narrowing
Summary:
The best fix for these narrowing
violations is to use static casting

Test Plan: verified

Reviewers: randall.rushing, Roach

Differential Revision: http://review.codefirex.com/D647

+++ camera/QCameraHWI_Parm.cpp
@@ -699,7 +699,7 @@ void QCameraHardwareInterface::loadTables()
 
 rat_t getRational(int num, int denom)
 {
-    rat_t temp = {num, denom};
+    rat_t temp = {static_cast<uint32_t>(num, denom)};
     return temp;
 }
